Dap.Karma: fix build (track recent API changes in master)
Recent master branch had an API break[1] to have long instead of int typed
row IDs for tracks. Adjust Dap.Karma extension to build again.

[1] https://git.gnome.org/browse/banshee/commit/?id=f847cb0c742b2d993dad71dba373bd81bb457afa

+++ src/Dap/Banshee.Dap.Karma/Banshee.Dap.Karma/KarmaSource.cs
@@ -47,8 +47,8 @@ namespace Banshee.Dap.Karma
         private String mount_point;
 
         // for getting karma track given a DatabaseTrackInfo
-        private Dictionary<int, KarmaTrackInfo> track_map =
-            new Dictionary<int, KarmaTrackInfo>();
+        private Dictionary<long, KarmaTrackInfo> track_map =
+            new Dictionary<long, KarmaTrackInfo>();
 
         public override void DeviceInitialize(IDevice dev)
         {
